nexusplan-backend/auth_service/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary
dj-database-url
drf-spectacular
djangorestframework-simplejwt
redis
python-dotenv
gunicorn